feat: set satellite bucket object expire to 14 days
Update the S3 module version so that it is possible to control if
delete markers are replicated to the destination bucket.

By doing this, and not enabling delete marker replication, it means
a shorter object expiration of 14 days can be set on the satellite
bucket objects which will save storage costs.

The central log archive bucket will keep its 90 day object expire
lifecycle rule so we'll still have the logs if needed.