feat: onboard LandingZone log-archive
Add the log-archive account so that we can use the existing
LandingZone CloudTrail S3 bucket to deliver CloudTrail
logs to the Central account.

The reason this account is being added is so the IAM policy
and roles are updated to allow replication from the LZ
log-archive bucket to the Central log-archive bucket.