feat: add a feature flag for the config rules
This is required as adding the config rules to the new Control Tower org
would mean having to go to the existing accounts goverened by CBS and
adding the control tower config rule recorder role.

Considering we are thinking of doing away with the config rules
completely and replacing them with either drift detection or a sentinel
alarm, we've decided that taking this step is an acceptable risk.