feat: remove all accounts non-prod, non-stage
All accounts that aren't in the Staging and Production OU will be
removed from the Cloud based sensor.

Control Tower has the ability to centralize CloudTrail Logs into a
single s3 bucket so we will set that up for CBS instead and all the rest
either aren't changed enough or don't use VPC's, ALB's, or WAFs in a way
that we care about.

This will be coupled with some manual removing of code.